added hot reloading
