+++ src/main/frontend/.postcssrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+    plugins: {
+        'autoprefixer': true
+    }
+}
