+++ src/main/frontend/package.json
@@ -25,7 +25,8 @@
     "popper.js": "^1.16.1",
     "query-string": "6.13.1",
     "vue": "2.6.12",
-    "vuetable-2": "^2.0.0-beta.4"
+    "vuetable-2": "^2.0.0-beta.4",
+    "webpack-jquery-ui": "^2.0.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
@@ -37,6 +38,7 @@
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-vue": "^6.2.2",
     "lint-staged": "^9.5.0",
+    "mini-css-extract-plugin": "^0.11.2",
     "prettier": "^1.19.1",
     "sass": "^1.26.5",
     "sass-loader": "^8.0.2",
