+++ src/main/frontend/src/components/ProjectList.vue
@@ -127,13 +127,12 @@
 
 <script>
 import Tag from "./Tag";
-import { clearFromEmpty } from "./../utils";
-import { Category, Platform, Visibility } from "../enums";
+import {clearFromEmpty, numberWithCommas} from "./../utils";
+import {Category, Platform, Visibility} from "../enums";
 import Pagination from "./Pagination";
 import Icon from "./Icon";
 import debounce from "lodash/debounce";
-import { API } from "../api";
-import { numberWithCommas } from "@/js/main";
+import {API} from "../api";
 
 export default {
   components: {
