+++ src/main/frontend/src/entrypoints/font-awesome.js
@@ -1,87 +1,88 @@
-import { config, dom, library } from "@fortawesome/fontawesome-svg-core";
+import {config, dom, library} from "@fortawesome/fontawesome-svg-core";
+import "@fortawesome/fontawesome-svg-core/styles.css";
 import {
-  faArrowLeft,
-  faArrowRight,
-  faAsterisk,
-  faBell,
-  faBook,
-  faBug,
-  faCalendar,
-  faChartArea,
-  faCheck,
-  faCheckCircle as fasCheckCircle,
-  faChevronDown,
-  faChevronUp,
-  faCircle,
-  faClipboard,
-  faCode,
-  faCog,
-  faComment,
-  faComments,
-  faDownload,
-  faEdit,
-  faExclamationCircle,
-  faExclamationTriangle,
-  faExternalLinkAlt,
-  faEye,
-  faEyeSlash,
-  faFlag,
-  faGamepad,
-  faGem as fasGem,
-  faGlobe,
-  faGraduationCap,
-  faHeartbeat,
-  faHome,
-  faInfo,
-  faInfoCircle,
-  faKey,
-  faLink,
-  faList,
-  faLock,
-  faMagic,
-  faMoneyBillAlt,
-  faPaperPlane,
-  faPencilAlt,
-  faPlay,
-  faPlayCircle,
-  faPlus,
-  faPuzzlePiece,
-  faQuestionCircle,
-  faReply,
-  faSave,
-  faSearch,
-  faServer,
-  faSignOutAlt,
-  faSpinner,
-  faStar as fasStar,
-  faStopCircle,
-  faTags,
-  faTerminal,
-  faThumbsUp as fasThumbsUp,
-  faTimes,
-  faTimesCircle,
-  faTrash,
-  faUnlockAlt,
-  faUpload,
-  faUser,
-  faUsers,
-  faUserTag,
-  faUserTie,
-  faWindowClose,
-  faWrench
+    faArrowLeft,
+    faArrowRight,
+    faAsterisk,
+    faBell,
+    faBook,
+    faBug,
+    faCalendar,
+    faChartArea,
+    faCheck,
+    faCheckCircle as fasCheckCircle,
+    faChevronDown,
+    faChevronUp,
+    faCircle,
+    faClipboard,
+    faCode,
+    faCog,
+    faComment,
+    faComments,
+    faDownload,
+    faEdit,
+    faExclamationCircle,
+    faExclamationTriangle,
+    faExternalLinkAlt,
+    faEye,
+    faEyeSlash,
+    faFlag,
+    faGamepad,
+    faGem as fasGem,
+    faGlobe,
+    faGraduationCap,
+    faHeartbeat,
+    faHome,
+    faInfo,
+    faInfoCircle,
+    faKey,
+    faLink,
+    faList,
+    faLock,
+    faMagic,
+    faMoneyBillAlt,
+    faPaperPlane,
+    faPencilAlt,
+    faPlay,
+    faPlayCircle,
+    faPlus,
+    faPuzzlePiece,
+    faQuestionCircle,
+    faReply,
+    faSave,
+    faSearch,
+    faServer,
+    faSignOutAlt,
+    faSpinner,
+    faStar as fasStar,
+    faStopCircle,
+    faTags,
+    faTerminal,
+    faThumbsUp as fasThumbsUp,
+    faTimes,
+    faTimesCircle,
+    faTrash,
+    faUnlockAlt,
+    faUpload,
+    faUser,
+    faUsers,
+    faUserTag,
+    faUserTie,
+    faWindowClose,
+    faWrench
 } from "@fortawesome/free-solid-svg-icons";
 
 import {
-  faCheckCircle as farCheckCircle,
-  faClock,
-  faFile,
-  faFileArchive,
-  faGem as farGem,
-  faMinusSquare,
-  faPlusSquare,
-  faSadTear,
-  faStar as farStar,
-  faThumbsUp as farThumbsUp
+    faCheckCircle as farCheckCircle,
+    faClock,
+    faFile,
+    faFileArchive,
+    faGem as farGem,
+    faMinusSquare,
+    faPlusSquare,
+    faSadTear,
+    faStar as farStar,
+    faThumbsUp as farThumbsUp
 } from "@fortawesome/free-regular-svg-icons";
 
 config.autoAddCss = false;
