--- src/main/frontend/src/entrypoints/main.js
@@ -1,1 +1,0 @@
-import "@/scss/main.scss";
