+++ src/main/frontend/src/js/main.js
@@ -1,10 +1,10 @@
 import $ from "jquery";
 import ClipboardJS from "clipboard";
 import hljs from "highlight.js";
+import {initTooltips, scrollToAnchor, toggleSpinner} from "@/utils";
+import "webpack-jquery-ui/tooltip";
 
 //=====> CONSTANTS
-
-export const KEY_ENTER = 13;
 
 //=====> SETUP
 
@@ -19,52 +19,6 @@ clipboardManager.on("success", function() {
 });
 
 //=====> HELPER FUNCTIONS
-
-export function sanitize(html) {
-  return $("<textarea>")
-    .html(html)
-    .text();
-}
-
-export function decodeHtml(html) {
-  // lol
-  return $("<textarea>")
-    .html(html)
-    .val();
-}
-
-export function go(str) {
-  window.location = decodeHtml(str);
-}
-
-export function clearUnread(e) {
-  e.find(".unread").remove();
-  if (!$(".user-dropdown .unread").length) $(".unread").remove();
-}
-
-function initTooltips() {
-  $('[data-toggle="tooltip"]').tooltip({
-    container: "body",
-    delay: { show: 500 }
-  });
-}
-
-export function slugify(name) {
-  return name
-    .trim()
-    .replace(/ +/g, " ")
-    .replace(/ /g, "-");
-}
-
-export function toggleSpinner(e) {
-  return e.toggleClass("fa-spinner").toggleClass("fa-spin");
-}
-
-export function numberWithCommas(x) {
-  const parts = x.toString().split(".");
-  parts[0] = parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ",");
-  return parts.join(".");
-}
 
 //=====> DOCUMENT READY
 
@@ -101,25 +55,6 @@ $(function() {
 
 // Fix page anchors which were broken by the fixed top navigation
 
-const scrollToAnchor = function(anchor) {
-  if (anchor) {
-    let target = $("a" + anchor);
-
-    if (target.length) {
-      $("html,body").animate(
-        {
-          scrollTop: target.offset().top - ($("#topbar").height() + 10)
-        },
-        1
-      );
-
-      return false;
-    }
-  }
-
-  return true;
-};
-
 $(window).on("load", function() {
   return scrollToAnchor(window.location.hash);
 });
@@ -137,7 +72,8 @@ $("a[href^='#']").click(function() {
 // The service worker has been removed in commit 9ab90b5f4a5728587fc08176e316edbe88dfce9e.
 // This code ensures that the service worker is removed from the browser.
 
-if (window.navigator && navigator.serviceWorker) {
+// Hangar doesn't need this since it was never added.
+/*if (window.navigator && navigator.serviceWorker) {
   if ("getRegistrations" in navigator.serviceWorker) {
     navigator.serviceWorker.getRegistrations().then(function(registrations) {
       registrations.forEach(function(registration) {
@@ -149,4 +85,4 @@ if (window.navigator && navigator.serviceWorker) {
       registration.unregister();
     });
   }
-}
+}*/
