+++ src/main/frontend/src/js/projectDiscuss.js
@@ -4,7 +4,7 @@
   var d = document.createElement("script");
   d.type = "text/javascript";
   d.async = true;
-  d.src = window.DiscourseEmbed.discourseUrl + "javascripts/embed.js";
+  d.src = window.DiscourseEmbed.discourseUrl + "build/js/embed.js"; // TODO this was ... + "javascripts/embed.js"
   (
     document.getElementsByTagName("head")[0] ||
     document.getElementsByTagName("body")[0]
