+++ src/main/frontend/src/js/showNotes.js
@@ -1,5 +1,5 @@
 import $ from "jquery";
-import { toggleSpinner } from "@/js/main";
+import {toggleSpinner} from "@/utils";
 
 //=====> DOCUMENT READY
