+++ src/main/java/freemarker_implicit.ftl
@@ -1,5 +1,6 @@
 [#ftl]
 [#-- @implicitly included --]
+[#-- @ftlvariable name="_csrf" type="org.springframework.security.web.csrf.CsrfToken" --]
 [#-- @ftlvariable name="cu" type="io.papermc.hangar.db.model.UsersTable" --]
 [#-- @ftlvariable name="markdownService" type="io.papermc.hangar.service.MarkdownService" --]
 [#-- @ftlvariable name="utils" type="io.papermc.hangar.util.TemplateHelper" --]
