+++ src/main/resources/templates/projects/settings.ftlh
@@ -13,12 +13,12 @@
         request: OreRequest[_], config: OreConfig, renderer: MarkdownRenderer, assetsFinder: AssetsFinder)-->
 
 <#assign scriptsVar>
-    <script type="text/javascript" src="<@hangar.url "javascripts/projectManage.js" />"></script>
-    <script type="text/javascript" src="<@hangar.url "javascripts/hideProject.js" />"></script>
-    <script type="text/javascript" src="<@hangar.url "javascripts/iconUpload.js" />"></script>
-    <script type="text/javascript" src="<@hangar.url "javascripts/keyGen.js" />"></script>
-    <script type="text/javascript" src="<@hangar.url "javascripts/userSearch.js" />"></script>
-    <script type="text/javascript" src="<@hangar.url "javascripts/memberList.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/projectManage.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/hideProject.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/iconUpload.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/keyGen.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/userSearch.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/memberList.js" />"></script>
     <script <#--@CSPNonce.attr -->>
         projectName = "${p.project.name}";
         PROJECT_OWNER = "${p.project.ownerName}";
