+++ src/main/resources/templates/swagger.ftlh
@@ -10,7 +10,7 @@ Mostly copied from swagger-ui's index file
     <script type="text/javascript" src="<@hangar.url "lib/jquery/dist/jquery.min.js" />"></script>
     <script src="<@hangar.url "lib/swagger-ui/swagger-ui-bundle.js" />"></script>
     <script src="<@hangar.url "lib/swagger-ui/swagger-ui-standalone-preset.js" />"></script>
-    <script src="<@hangar.url "javascripts/apiRequests.js" />"></script>
+    <script src="<@hangar.url "build/js/apiRequests.js" />"></script>
     <script @CSPNonce.attr>
         window.onload = () => {
             window.ui = SwaggerUIBundle({
