+++ src/main/resources/templates/users/admin/visibility.ftlh
@@ -39,7 +39,7 @@
 </#macro>
 
 <#assign scriptsVar>
-    <script type="text/javascript" src="<@hangar.url "javascripts/hideProject.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/hideProject.js" />"></script>
 </#assign>
 
 <@base.base title="Project Visibility" additionalScripts=scriptsVar>
