+++ src/main/resources/templates/users/apiKeys.ftlh
@@ -4,7 +4,7 @@
 
 <#--@(u: UserData, o: Option[(OrganizationData, ScopedOrganizationData)], keys: Seq[ApiKey], perms: Seq[NamedPermission])(implicit messages: Messages, flash: Flash, request: OreRequest[_], config: OreConfig, assetsFinder: AssetsFinder)-->
 <#assign scriptsVar>
-    <script type="text/javascript" src="<@hangar.url "javascripts/apiKeysManagement.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/apiKeysManagement.js" />"></script>
     <script <#--@CSPNonce.attr-->>
         NO_PERMS_SET = '<@spring.message "user.apiKeys.error.noPermsSet" />';
         NO_NAME_SET = '<@spring.message "user.apiKeys.error.noNameSet" />';
