+++ src/main/resources/templates/users/projects.ftlh
@@ -20,12 +20,12 @@
 </#function>
 
 <#assign scriptsVar>
-    <script type="text/javascript" src="<@hangar.url "build/user-profile.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/user-profile.js" />"></script>
     <#if u.isOrga() && o?? && canEditOrgMembers()>
-    <script type="text/javascript" src="<@hangar.url "javascripts/orgInvites.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/orgInvites.js" />"></script>
     </#if>
-    <script type="text/javascript" src="<@hangar.url "javascripts/userSearch.js" />"></script>
-    <script type="text/javascript" src="<@hangar.url "javascripts/memberList.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/userSearch.js" />"></script>
+    <script type="text/javascript" src="<@hangar.url "build/js/memberList.js" />"></script>
 </#assign>
 
 <@users.view u=u o=o additionalScripts=scriptsVar>
