+++ src/main/resources/templates/utils/hangar.ftlh
@@ -5,13 +5,16 @@
 
 <#macro url path>
     <#compress>
-<#--        <#if path == 'build/main.css'>-->
-<#--            http://localhost:8081/main.css-->
-<#--        <#else>-->
-            <#if path?starts_with("bootstrap") || path?starts_with("build") || path?starts_with("diff-match-patch") || path?starts_with("highlight") || path?starts_with("images") || path?starts_with("javascripts") || path?starts_with("manifest") || path?starts_with('jquery-ui')>
+        <#if path?starts_with("build") && config.debug>
+            http://localhost:8081/${path[6..]}
+        <#elseif path?starts_with("bootstrap") || path?starts_with("build") || path?starts_with("diff-match-patch") || path?starts_with("highlight") || path?starts_with("images") || path?starts_with("javascripts") || path?starts_with("manifest") || path?starts_with('jquery-ui')>
+            /assets/${path}
+        <#else>
+            /${path}
+        </#if>
+<#--            <#if path?starts_with("bootstrap") || path?starts_with("build") || path?starts_with("diff-match-patch") || path?starts_with("highlight") || path?starts_with("images") || path?starts_with("javascripts") || path?starts_with("manifest") || path?starts_with('jquery-ui')>
             /assets
             </#if>
-            /${path}
-<#--        </#if>-->
+            /${path}-->
     </#compress>
 </#macro>
